Refactor project to have graphql in its own file
