+++ package.json
@@ -2,7 +2,7 @@
   "name": "msg",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "server/server.js",
   "dependencies": {
     "acorn": "^5.1.1",
     "astring": "^1.0.3",
@@ -19,9 +19,9 @@
   },
   "devDependencies": {},
   "scripts": {
-    "build": "node src/setup.js",
+    "build": "node server/setup.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node build/server.js"
+    "start": "node server/server.js"
   },
   "author": "",
   "license": "ISC"
