+++ server/constants.js
@@ -1,0 +1,7 @@
+const endpoint = 'https://graph.microsoft.com/';
+const version = 'v1.0';
+const graphNameSpace = 'microsoft.graph.';
+
+const graphRoot = endpoint + version;
+
+module.exports = { endpoint, version, graphRoot, graphNameSpace };
